Fix a bug in the database callback code.  Moved the actual callback function calls into the item object where they belong.
The signature for a callback was just the item key and the name
of the plugin.  This only allowed one callback per item per key.
The netfix plugin needs to add a different set of callbacks for each
client connection.  The functions for each connection are unique
so they are used in the signature of the callback along with the udata
parameter.  Also the callback calling code was moved to the db_item
class so that it'll get called no matter how the value or aux data is
changed.  Will need to so something similar for all of the flags.